+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Converter for Lithuania (EuroCrops)
 - Converter for Switzerland
 - Converter for Slovenia
+- Converter for Slovakia
 
 ### Fixed